requirements.txt
chumpy==0.70
datasets==2.18.0
einops==0.8.0
face-alignment==1.4.1
imageio==2.34.0
imageio-ffmpeg==0.4.9
ipdb==0.13.13
kornia==0.7.2
loguru==0.7.2
lpips==0.1.4
matplotlib==3.8.4
numpy==1.26.4
opencv-python==4.9.0.80
openmesh==1.2.1
Pillow==10.3.0
pytorch-msssim==1.0.0
PyYAML==6.0.1
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.11.4
setuptools>=68.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
yacs==0.1.8